moved logics to controller file
